Fix exim not using faketime
This happened because the subprocesses were not getting the `LD_PRELOAD`
env variable. But also because the `LD_PRELOAD` was never applied due to
a Linux safety guard. Basically the exim binary had a setuid bit which
caused it to be loaded with `root` permissions while the actual user is
`Debian-exim`. This mismatch between the ruid and euid meant that the
`LD_PRELOAD` was never loaded.