Control time via shared timestamp file
Identical runs can have small exim timing differences. This would cause
them to appear different (aka not have matching diffs because of
timestamps). This controls the time in the exim containers better by
making the time fixed. It then uses a file to be able to advance the
time in a controlled manner.